pyproject.toml
[project]
name = "pelican-social-cards"
version = "0.5.0"
description = "Plugin to generate social media cards with post title embedded"
authors = [
    {name = "Mirek Długosz", email = "mirek@mirekdlugosz.com"},
]
license = { text = "AGPL-3.0" }
readme = "README.md"
keywords = ["pelican", "plugin"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: Pelican",
    "Framework :: Pelican :: Plugins",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "pelican>=4.5",
    "pillow (>=11.1,<12.0)",
]

[project.optional-dependencies]
markdown = ["markdown>=3.4"]
lint = [
    "invoke>=2.2",
    "ruff>=0.9.1,<1.0.0",
]
test = [
    "markdown>=3.4",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-sugar>=1.0",
]

[project.urls]
Homepage = "https://github.com/mirekdlugosz/pelican-social-cards"
Repository = "https://github.com/mirekdlugosz/pelican-social-cards"

[build-system]
requires = ["poetry>=2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "pelican" },
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"

[tool.ruff.lint]
select = [
  "B",   # flake8-bugbear
  "BLE", # flake8-blind-except
  "C4",  # flake8-comprehensions
  "D",   # pydocstyle
  "E",   # pycodestyle
  "F",   # pyflakes
  "I",   # isort
  "ICN", # flake8-import-conventions
  "ISC", # flake8-implicit-str-concat
  "PGH", # pygrep-hooks
  "PL",  # pylint
  "RET", # flake8-return
  "RUF", # ruff-specific rules
  "SIM", # flake8-simplify
  "T10", # flake8-debugger
  "T20", # flake8-print
  "TID", # flake8-tidy-imports
  "TRY", # tryceratops
  "UP",  # pyupgrade
  "W",   # pycodestyle
  "YTT", # flake8-2020
]

ignore = [
  "D100",    # missing docstring in public module
  "D104",    # missing docstring in public package
  "D203",    # blank line before class docstring
  "D213",    # multi-line docstring summary should start at the second line
  "ISC001",  # disabled so `ruff format` works without warning
  "RET504",  # Unnecessary assignment before `return` statement
  # Not shared by other plugins
  "D103",    # Missing docstring in public function
  "D107",    # Missing docstring in `__init__`
]

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true
known-first-party = ["pelican"]
